+++ common/queuedetailspanel.h
@@ -73,8 +73,8 @@ public slots:
         void setGuiOptions(const QVariant &);
         void setUserInfo(const UserInfo *);
 	void newQueue(const QString &, const QString &, const QVariant &);
-        void setAgentList(int, const QVariant &);
-        void updatePeerAgent(int,
+        void setAgentList(double, const QVariant &);
+        void updatePeerAgent(double,
                              const QString &,
                              const QString &,
                              const QVariant &);
