+++ common/switchboardwindow.h
@@ -81,7 +81,7 @@ public slots:
                         const QString &,
                         const QVariant &);
         void newUser(UserInfo *);
-        void updatePeerAgent(int,
+        void updatePeerAgent(double,
                              const QString &,
                              const QString &,
                              const QVariant &);
